Performance optimisations and bugfixes - breaking changes
There was some questionable decisions with component evaluation - I overcomplicated the problem in my initial attempt and it went through many issues. Now I've solved the problem in a quicker and more optimised way by leveraging template strings. This is why this is a breaking change, all  needs to be updated to . This commit also fixes all the examples with this change. But there are clear gains for this. One is that functions that have multiple  characters now work - previously they didn't but it hadn't come up in testing. Now it is possible to write full functions in . And, since they are evaluated as template strings, the process for evaluating them is simpler and quicker. There were also some easy performance gains after researching  and  differences for clearing all tags. There are still areas (like comment filtering) that could be optimised for *rerendering only* (it adds time to component creation).