moved heavy operations during rerender to component creation
Originally the idea was that all `<script>` logic would be run \"at runtime\", meaning most of the logic was contained within `rerender()` - after the previous comment filtering caches I looked at the `rerender()` function and realised that most of the startup-related code could go on connectedCallback and could allow me to cache the filter results. This means that `rerender()` no longer a) runs `\<style\>`s and `\<script\>`s, or b) filters scripts. I believe the original reason for putting everything in rerender was to do with some issues relating to components needing to be upgraded and getting fallback to work (fallback was the default, and it was immediately rendered - then swapped *after* the fetch... at some point this changed to using a helper `define()` function that would take the fallback as an argument, so the original reason for everything being structured like this was no longer valid)